# Make AntiKt.PseudoJet and HistoryElement immutable