--- a/antiktjl/Project.toml
+++ b/antiktjl/Project.toml
@@ -2,3 +2,6 @@ name = "AntiKt"
 uuid = "b1d79ec0-5126-4cc0-8710-7ce7530fd203"
 authors = ["Philippe Gras <philippe.gras@cern.ch>"]
 version = "0.1.0"
+
+[deps]
+Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"