--- a/antiktjl/src/AntiKt.jl
+++ b/antiktjl/src/AntiKt.jl
@@ -28,6 +28,8 @@
 """
 module AntiKt
 
+using Accessors
+
 export PseudoJet, antikt
 
 include("pseudojet.jl")