This turns `AntiKt.PseudoJet` and `HistoryElement` immutable. It doesn't speed up the algorithm itself, but reduces memory allocation significantly:

Before:

```julia
julia> @benchmark antikt.(events, R, ptmin)
BenchmarkTools.Trial: 160 samples with 1 evaluation.
 Range (min … max):  29.220 ms … 35.208 ms  ┊ GC (min … max): 0.00% … 11.49%
 Time  (median):     30.464 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   31.331 ms ±  1.620 ms  ┊ GC (mean ± σ):  4.11% ±  4.71%
```

After:

```julia
BenchmarkTools.Trial: 159 samples with 1 evaluation.
 Range (min … max):  30.142 ms … 37.150 ms  ┊ GC (min … max): 0.00% … 3.74%
 Time  (median):     31.688 ms              ┊ GC (median):    3.63%
 Time  (mean ± σ):   31.560 ms ±  1.001 ms  ┊ GC (mean ± σ):  2.02% ± 1.96%
```

The memory allocation cost of mutable structs is higher in Julia than in C++: An array of fixed-sized mutable structs is only a single memory allocation in C++, but in Julia, every element of the array must be allocated on the heap separately (this is true for pretty much any language with automatic garbage collection). "Small" structs should be immutable in Julia.

This becomes especially relevant when running on many threads, since threads can't do heap-(de-)allocation in parallel (both in Julia and C++, also Julia doesn't to parallel GC yet), so the alloc/gc-percentage goes up as the number of threads increases, until alloc/GC-cost dominates.

The remaining alloc/GC cost in `antikt` is mainly due to the mutable `TiledJet`. Using indices instead of pointers for `TiledJet`-references would make it possible to make `TiledJet` immutable as well, reducing GC cost to almost zero. `ArraysOfArrays.VectorOfArrays` could be used to reduce memory allocations for the vector-of-vectors-of-jets that `antikt` returns.

In general, using graph data structures based on adjacency lists (e.g. via Graphs.jl) would probably be better than using pointer/reference-based graph structures.