@oschulz 2022-11-08
> By making PseudoJet immutable you have dropped the on-demand calculation of rapidity and phi. It results in a 10% increase of runtime.

Yes, I did, though in my benchmarks I didn't see a significant performance penalty due to that. But then, effects on the order of 10% are tricky to benchmark because they can depend on the data size, caching effects, multi- vs. single-threaded execution, etc.

The rapidity and phi calculation looked fairly cheap, so I though it might be easiest to not cache them. In principle we could implement a caching for them while keeping `PseudoJet` immutable, I think, but it may not  be worth it (would require more in-depth benchmarking) - it could even lead to worse performance in some scenarios due to increased branch-prediction misses in the CPU. It will also depend on how often these values are requested for the same jet, and for what percentage of jets they don't need to be calculated at all, of course.


> Is your point of individual field allocation also true for struct containig only PODs

Julia (pretty much any language with automatic GC) needs to allocate any mutable struct on the heap. Immutable structs that have mutable structs as field values can actually be stack allocated now (since v1.6, I think). The reason is that mutable structs are "unique individuals" - if you have a vector of them, some part of the code may hold a reference to one of the elements. So the vector can't be allocated/deallocated as a whole, each element needs to be allocated and GCed individually, in addition to the vector itself (which doesn't contain the mutable PODs, only pointers to them).

Struct *fields* have to be heap allocated if they are mutable or not concretely typed (e.g. a field of type `AbstractFloat`), whereas fields of and immutable concrete data type are stack-allocated, "embedded" in the object that contains them.